in paip_analysis_complete
more work on creating multi plots per image

weights added to plots in do_REM function

These changes have not been rolled out to all scripts. There is still work to do in tidying up the scripts to have a consistent naming convention and structure. Also, should remove models 3&4, to be run only if model 2 is significan. Create an example of stratified analysis to be used if models 5-8 are significant.